+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 - none
 
 # v2024.11.1-beta
+- sqlmath - Streamline idate-sql-functions to just:
+        idateadd()
+        idatefrom()
+        idateto()
+        idateymdfrom()
 - sqlmath - Add sql-functions IDATETOIDATE(), IDATETOIDATETIME().
 - sqlmath - Add sql-functions IDATEFROMEPOCH(), IDATETIMEFROMEPOCH().
 - sqlmath - Add sql-function IDATEADD().